COMPMID-417 Fix bare metal build for armv7a
Reviewed-on: http://mpd-gerrit.cambridge.arm.com/89670
Tested-by: Kaizen <jeremy.johnson+kaizengerrit@arm.com>

+++ src/core/NEON/kernels/NEFillBorderKernel.cpp
@@ -99,7 +99,7 @@ class Coordinates;
 } // namespace arm_compute
 
 NEFillBorderKernel::NEFillBorderKernel()
-    : _tensor(nullptr), _border_size(0), _mode(BorderMode::UNDEFINED), _constant_border_value(0)
+    : _tensor(nullptr), _border_size(0), _mode(BorderMode::UNDEFINED), _constant_border_value(static_cast<float>(0.f))
 {
 }
