+++ src/runtime/NEON/functions/NEDepthConcatenate.cpp
@@ -52,7 +52,7 @@ void NEDepthConcatenate::configure(std::vector<ITensor *> inputs_vector, ITensor
     for(unsigned int i = 0; i < _num_inputs; ++i)
     {
         _concat_kernels_vector[i].configure(inputs_vector.at(i), depth_offset, output);
-        _border_handlers_vector[i].configure(inputs_vector.at(i), _concat_kernels_vector[i].border_size(), BorderMode::CONSTANT, PixelValue(0));
+        _border_handlers_vector[i].configure(inputs_vector.at(i), _concat_kernels_vector[i].border_size(), BorderMode::CONSTANT, PixelValue(static_cast<float>(0.f)));
 
         depth_offset += inputs_vector.at(i)->info()->dimension(2);
     }
